requirements.txt
torch
transformers
Pillow
einops
torchvision
huggingface_hub
bitsandbytes
accelerate
numpy
tensorflow
